+++ src/IECoreRenderMan/Attributes.h
@@ -63,6 +63,8 @@ class Attributes : public IECoreScenePreview::Renderer::AttributesInterface
 		const RtParamList &prototypeAttributes() const;
 		/// Attributes to be applied to GeometryInstances.
 		const RtParamList &instanceAttributes() const;
+		/// Attributes to be applied to light GeometryInstances.
+		const RtParamList &lightInstanceAttributes() const;
 
 		const Material *material() const;
 		const Displacement *displacement() const { return m_displacement.get(); }
@@ -75,16 +77,20 @@ class Attributes : public IECoreScenePreview::Renderer::AttributesInterface
 
 		const IECoreScene::ShaderNetwork *lightFilter() const;
 
+		const bool isUSDMeshLight() const;
+
 	private :
 
 		std::optional<IECore::MurmurHash> m_prototypeHash;
 		RtParamList m_prototypeAttributes;
 		RtParamList m_instanceAttributes;
+		std::optional<RtParamList> m_lightInstanceAttributes;
 		ConstMaterialPtr m_material;
 		ConstDisplacementPtr m_displacement;
 		IECoreScene::ConstShaderNetworkPtr m_lightShader;
 		ConstMaterialPtr m_lightMaterial;
 		IECoreScene::ConstShaderNetworkPtr m_lightFilter;
+		bool m_isUSDMeshLight;
 
 };