+++ src/IECoreRenderManModule/IECoreRenderManModule.cpp
@@ -36,6 +36,8 @@
 
 #include "boost/python.hpp"
 
+#include "IECorePython/ScopedGILRelease.h"
+
 #include "IECoreRenderMan/ShaderNetworkAlgo.h"
 
 #include "prmanapi.h"
@@ -70,6 +72,13 @@ ShaderNetworkAlgo::VStructAction evaluateVStructConditionalWrapper( const std::s
 	);
 }
 
+IECore::CompoundObjectPtr convertUSDMeshLightAttributesWrapper( const IECore::CompoundObject &attributes, bool copy )
+{
+	IECorePython::ScopedGILRelease r;
+	IECore::ConstCompoundObjectPtr result = ShaderNetworkAlgo::convertUSDMeshLightAttributes( &attributes );
+	return copy ? result->copy() : boost::const_pointer_cast<IECore::CompoundObject>( result );
+}
+
 } // namespace
 
 BOOST_PYTHON_MODULE( _IECoreRenderMan )
@@ -84,6 +93,8 @@ BOOST_PYTHON_MODULE( _IECoreRenderMan )
 
 	def( "convertUSDShaders", &ShaderNetworkAlgo::convertUSDShaders );
 	def( "usdLightTransform", &ShaderNetworkAlgo::usdLightTransform );
+
+	def( "convertUSDMeshLightAttributes", &convertUSDMeshLightAttributesWrapper, ( arg_( "_copy" ) = true ) );
 
 	{
 		scope s = class_<ShaderNetworkAlgo::VStructAction>( "VStructAction" )